fix: authorization to delete an attendance that is not from the patient in request

+++ src/data/services/attendance/db-delete-attendance.ts
@@ -1,21 +1,33 @@
-import { DeleteAttendance } from "../../../domain/features/attendance/delete-attendance";
-import { DeleteAttendanceRepository } from "../../contracts/repos/attendance-repository";
-import { GetOnePatient } from "../../../domain/features/patient/get-one-patient";
+import { DeleteAttendance, GetOnePatient } from "@/domain/features";
+import { NotAuthorizedException, NotFoundException } from '@/presentation/errors';
+import {
+	GetOneAttendanceRepository,
+	DeleteAttendanceRepository,
+} from "@/data/contracts/repos";
 
 export class DbDeleteAttendance implements DeleteAttendance {
 	constructor(
-		private readonly repository: DeleteAttendanceRepository,
+		private readonly repository: DeleteAttendanceRepository & GetOneAttendanceRepository,
 		private readonly getOnePatient: GetOnePatient
 	) {}
 
 	async delete(
 		params: DeleteAttendance.Params
 	): Promise<DeleteAttendance.Result> {
+
+			
 		const patient = await this.getOnePatient.getOne({
 			patientId: params.patientId,
 			userId: params.userId,
 		});
 		if (patient instanceof Error) return patient;
+		const attendance = await this.repository.getOne({
+			id: params.attendanceId
+		})	
+		if(attendance && attendance.patientId !== params.patientId) {
+			return new NotFoundException('Attendance')	
+		}	
+			
 		return await this.repository.delete({
 			id: params.attendanceId,
 		});
